extract milestone logic
