+++ checkpoint/handler_milestone_test.go
@@ -1,0 +1,211 @@
+package checkpoint_test
+
+import (
+	sdk "github.com/cosmos/cosmos-sdk/types"
+
+	chSim "github.com/maticnetwork/heimdall/checkpoint/simulation"
+	"github.com/maticnetwork/heimdall/checkpoint/types"
+	errs "github.com/maticnetwork/heimdall/common"
+	"github.com/maticnetwork/heimdall/helper"
+
+	"github.com/stretchr/testify/require"
+
+	hmTypes "github.com/maticnetwork/heimdall/types"
+)
+
+func (suite *HandlerTestSuite) TestHandleMsgMilestone() {
+	t, app, ctx := suite.T(), suite.app, suite.ctx
+	keeper := app.CheckpointKeeper
+	stakingKeeper := app.StakingKeeper
+	start := uint64(0)
+	borChainId := "1234"
+	milestoneID := "0000"
+	milestoneLength := helper.MilestoneLength
+
+	// check valid milestone
+	// generate proposer for validator set
+	chSim.LoadValidatorSet(t, 2, stakingKeeper, ctx, false, 10)
+	stakingKeeper.IncrementAccum(ctx, 1)
+
+	lastMilestone, err := keeper.GetLastMilestone(ctx)
+	if err == nil {
+		start = start + lastMilestone.EndBlock + 1
+	}
+
+	header, err := chSim.GenRandMilestone(start, milestoneLength)
+	require.NoError(t, err)
+
+	// add current proposer to header
+	header.Proposer = stakingKeeper.GetValidatorSet(ctx).Proposer.Signer
+
+	suite.Run("Success", func() {
+		msgMilestone := types.NewMsgMilestoneBlock(
+			header.Proposer,
+			header.StartBlock,
+			header.EndBlock,
+			header.RootHash,
+			borChainId,
+			milestoneID,
+		)
+
+		// send milestone to handler
+		got := suite.handler(ctx, msgMilestone)
+		require.True(t, got.IsOK(), "expected send-milstone to be ok, got %v", got)
+		bufferedHeader, _ := keeper.GetLastMilestone(ctx)
+		require.Empty(t, bufferedHeader, "Should not store state")
+	})
+
+	suite.Run("Invalid msg based on sprint length", func() {
+		header.Proposer = hmTypes.HexToHeimdallAddress("1234")
+		msgMilestone := types.NewMsgMilestoneBlock(
+			header.Proposer,
+			header.StartBlock,
+			header.EndBlock,
+			header.RootHash,
+			borChainId,
+			milestoneID,
+		)
+
+		// send milestone to handler
+		got := suite.handler(ctx, msgMilestone)
+		require.True(t, !got.IsOK(), errs.CodeToDefaultMsg(got.Code))
+		require.Equal(t, errs.CodeInvalidMsg, got.Code)
+	})
+
+	suite.Run("Invalid Proposer", func() {
+		header.Proposer = hmTypes.HexToHeimdallAddress("1234")
+		msgMilestone := types.NewMsgMilestoneBlock(
+			header.Proposer,
+			header.StartBlock,
+			header.EndBlock,
+			header.RootHash,
+			borChainId,
+			milestoneID,
+		)
+
+		// send milestone to handler
+		got := suite.handler(ctx, msgMilestone)
+		require.True(t, !got.IsOK(), errs.CodeToDefaultMsg(got.Code))
+		require.Equal(t, errs.CodeInvalidMsg, got.Code)
+	})
+
+	suite.Run("Milestone not in countinuity", func() {
+
+		err = keeper.AddMilestone(ctx, header)
+		require.NoError(t, err)
+
+		_, err = keeper.GetLastMilestone(ctx)
+		require.NoError(t, err)
+
+		lastMilestone, err := keeper.GetLastMilestone(ctx)
+		if err == nil {
+			// pass wrong start
+			start = start + lastMilestone.EndBlock + 2
+		}
+
+		msgMilestone := types.NewMsgMilestoneBlock(
+			header.Proposer,
+			start,
+			start+milestoneLength-1,
+			header.RootHash,
+			borChainId,
+			milestoneID,
+		)
+
+		// send milestone to handler
+		got := suite.handler(ctx, msgMilestone)
+		require.True(t, !got.IsOK(), errs.CodeToDefaultMsg(got.Code))
+		require.Equal(t, errs.CodeMilestoneNotInContinuity, got.Code)
+
+	})
+
+	suite.Run("Milestone not in countinuity", func() {
+
+		_, err = keeper.GetLastMilestone(ctx)
+		require.NoError(t, err)
+
+		lastMilestone, err := keeper.GetLastMilestone(ctx)
+		if err == nil {
+			// pass wrong start
+			start = start + lastMilestone.EndBlock - 2
+		}
+
+		msgMilestone := types.NewMsgMilestoneBlock(
+			header.Proposer,
+			start,
+			start+milestoneLength-1,
+			header.RootHash,
+			borChainId,
+			milestoneID,
+		)
+
+		// send milestone to handler
+		got := suite.handler(ctx, msgMilestone)
+		require.True(t, !got.IsOK(), errs.CodeToDefaultMsg(got.Code))
+		require.Equal(t, errs.CodeMilestoneNotInContinuity, got.Code)
+
+	})
+
+}
+
+func (suite *HandlerTestSuite) TestHandleMsgMilestoneExistInStore() {
+	t, app, ctx := suite.T(), suite.app, suite.ctx
+	keeper := app.CheckpointKeeper
+	stakingKeeper := app.StakingKeeper
+	start := uint64(0)
+	milestoneLength := helper.MilestoneLength
+
+	chSim.LoadValidatorSet(t, 2, stakingKeeper, ctx, false, 10)
+	stakingKeeper.IncrementAccum(ctx, 1)
+
+	lastMilestone, err := keeper.GetLastMilestone(ctx)
+	if err == nil {
+		start = start + lastMilestone.EndBlock + 1
+	}
+
+	header, err := chSim.GenRandMilestone(start, milestoneLength)
+	require.NoError(t, err)
+
+	// add current proposer to header
+	header.Proposer = stakingKeeper.GetValidatorSet(ctx).Proposer.Signer
+
+	// send old milestone
+	res := suite.SendMilestone(header)
+
+	require.True(t, res.IsOK(), "expected send-milestone to be  ok, got %v", res)
+
+	// send milestone to handler
+	got := suite.SendMilestone(header)
+	require.True(t, !got.IsOK(), errs.CodeToDefaultMsg(got.Code))
+}
+
+func (suite *HandlerTestSuite) SendMilestone(header hmTypes.Milestone) (res sdk.Result) {
+	_, ctx := suite.app, suite.ctx
+
+	milestoneLength := helper.MilestoneLength
+
+	// keeper := app.MilestoneKeeper
+
+	borChainId := "1234"
+	milestoneID := "00000"
+	// create milestone msg
+	msgMilestone := types.NewMsgMilestoneBlock(
+		header.Proposer,
+		header.StartBlock,
+		header.EndBlock,
+		header.RootHash,
+		borChainId,
+		milestoneID,
+	)
+
+	suite.contractCaller.On("CheckIfBlocksExist", header.EndBlock).Return(true)
+	suite.contractCaller.On("GetRootHash", header.StartBlock, header.EndBlock, milestoneLength).Return(header.RootHash.Bytes(), nil)
+	suite.contractCaller.On("GetVoteOnRootHash", header.StartBlock, header.EndBlock, milestoneLength, header.RootHash.String(), header.MilestoneID).Return(true, nil)
+
+	// send milestone to handler
+	result := suite.handler(ctx, msgMilestone)
+	sideResult := suite.sideHandler(ctx, msgMilestone)
+	suite.postHandler(ctx, msgMilestone, sideResult.Result)
+
+	return result
+}
