+++ cmd/heimdalld/haimdalld/service.go
@@ -1,0 +1,472 @@
+package haimdalld
+
+import (
+	"context"
+	"encoding/hex"
+	"encoding/json"
+	"fmt"
+	"os"
+	"os/signal"
+	"path/filepath"
+	"runtime/pprof"
+	"syscall"
+	"time"
+
+	"github.com/cosmos/cosmos-sdk/baseapp"
+	"github.com/cosmos/cosmos-sdk/client/flags"
+	"github.com/cosmos/cosmos-sdk/server"
+	"github.com/cosmos/cosmos-sdk/store"
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	jsoniter "github.com/json-iterator/go"
+	"github.com/spf13/cobra"
+	"github.com/spf13/viper"
+	dbm "github.com/tendermint/tm-db"
+	"go.opentelemetry.io/otel"
+	"go.opentelemetry.io/otel/exporters/otlp/otlptrace"
+	"go.opentelemetry.io/otel/exporters/otlp/otlptrace/otlptracegrpc"
+	"go.opentelemetry.io/otel/propagation"
+	"go.opentelemetry.io/otel/sdk/resource"
+	sdktrace "go.opentelemetry.io/otel/sdk/trace"
+	"golang.org/x/sync/errgroup"
+	"google.golang.org/grpc"
+
+	"github.com/maticnetwork/heimdall/app"
+	bridgeCmd "github.com/maticnetwork/heimdall/bridge/cmd"
+	"github.com/maticnetwork/heimdall/helper"
+	restServer "github.com/maticnetwork/heimdall/server"
+)
+
+var logger = helper.Logger.With("module", "cmd/heimdalld")
+
+func New() {
+	cdc := app.MakeCodec()
+	ctx := server.NewDefaultContext()
+
+	shutdownCtx, stop := signal.NotifyContext(context.Background(), os.Interrupt, syscall.SIGTERM)
+	defer stop()
+
+	rootCmd := &cobra.Command{
+		Use:               "heimdalld",
+		Short:             "Heimdall Daemon (server)",
+		PersistentPreRunE: server.PersistentPreRunEFn(ctx),
+	}
+
+	// adding heimdall configuration flags to root command
+	helper.DecorateWithHeimdallFlags(rootCmd, viper.GetViper(), logger, "main")
+	helper.DecorateWithTendermintFlags(rootCmd, viper.GetViper(), logger, "main")
+
+	tendermintCmd := &cobra.Command{
+		Use:   "tendermint",
+		Short: "Tendermint subcommands",
+	}
+
+	rootCmd.AddCommand(heimdallStart(shutdownCtx, ctx, getNewApp(ctx), cdc)) // New Heimdall start command
+
+	tendermintCmd.AddCommand(
+		server.ShowNodeIDCmd(ctx),
+		server.ShowValidatorCmd(ctx),
+		server.ShowAddressCmd(ctx),
+		server.VersionCmd(ctx),
+	)
+
+	rootCmd.AddCommand(server.UnsafeResetAllCmd(ctx))
+	rootCmd.AddCommand(flags.LineBreak)
+	rootCmd.AddCommand(tendermintCmd)
+	rootCmd.AddCommand(server.ExportCmd(ctx, cdc, exportAppStateAndTMValidators))
+	rootCmd.AddCommand(flags.LineBreak)
+	rootCmd.AddCommand(version.Cmd) // Using heimdall version, not Cosmos SDK version
+	// End of block
+
+	rootCmd.AddCommand(showAccountCmd())
+	rootCmd.AddCommand(showPrivateKeyCmd())
+	rootCmd.AddCommand(restServer.ServeCommands(shutdownCtx, cdc, restServer.RegisterRoutes))
+	rootCmd.AddCommand(bridgeCmd.BridgeCommands(viper.GetViper(), logger, "main"))
+	rootCmd.AddCommand(VerifyGenesis(ctx, cdc))
+	rootCmd.AddCommand(initCmd(ctx, cdc))
+	rootCmd.AddCommand(testnetCmd(ctx, cdc))
+
+	// rollback cmd
+	rootCmd.AddCommand(rollbackCmd(ctx))
+
+	// prepare and add flags
+	executor := cli.PrepareBaseCmd(rootCmd, "HD", os.ExpandEnv("/var/lib/heimdall"))
+	if err := executor.Execute(); err != nil {
+		// Note: Handle with #870
+		panic(err)
+	}
+}
+
+func getNewApp(serverCtx *server.Context) func(logger log.Logger, db dbm.DB, storeTracer io.Writer) abci.Application {
+	return func(logger log.Logger, db dbm.DB, storeTracer io.Writer) abci.Application {
+		// init heimdall config
+		helper.InitHeimdallConfig("")
+		helper.UpdateTendermintConfig(serverCtx.Config, viper.GetViper())
+		// create new heimdall app
+		return app.NewHeimdallApp(logger, db, baseapp.SetPruning(store.NewPruningOptionsFromString(viper.GetString("pruning"))))
+	}
+}
+
+func exportAppStateAndTMValidators(logger log.Logger, db dbm.DB, storeTracer io.Writer, height int64, forZeroHeight bool, jailWhiteList []string) (json.RawMessage, []tmTypes.GenesisValidator, error) {
+	bapp := app.NewHeimdallApp(logger, db)
+	return bapp.ExportAppStateAndValidators()
+}
+
+func heimdallStart(shutdownCtx context.Context, ctx *server.Context, appCreator server.AppCreator, cdc *codec.Codec) *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "start",
+		Short: "Run the full node",
+		Long: `Run the full node application with Tendermint in process.
+Starting rest server is provided with the flag --rest-server and starting bridge with
+the flag --bridge when starting Tendermint in process.
+Pruning options can be provided via the '--pruning' flag. The options are as follows:
+
+syncable: only those states not needed for state syncing will be deleted (keeps last 100 + every 10000th)
+nothing: all historic states will be saved, nothing will be deleted (i.e. archiving node)
+everything: all saved states will be deleted, storing only the current state
+
+Node halting configurations exist in the form of two flags: '--halt-height' and '--halt-time'. During
+the ABCI Commit phase, the node will check if the current block height is greater than or equal to
+the halt-height or if the current block time is greater than or equal to the halt-time. If so, the
+node will attempt to gracefully shutdown and the block will not be committed. In addition, the node
+will not be able to commit subsequent blocks.
+
+For profiling and benchmarking purposes, CPU profiling can be enabled via the '--cpu-profile' flag
+which accepts a path for the resulting pprof file.
+`,
+		RunE: func(cmd *cobra.Command, args []string) error {
+			ctx.Logger.Info("starting ABCI with Tendermint")
+
+			startRestServer, _ := cmd.Flags().GetBool(helper.RestServerFlag)
+			startBridge, _ := cmd.Flags().GetBool(helper.BridgeFlag)
+
+			err := startInProcess(cmd, shutdownCtx, ctx, appCreator, cdc, startRestServer, startBridge)
+			return err
+		},
+		PreRun: func(cmd *cobra.Command, args []string) {
+			// bridge binding
+			if err := viper.BindPFlag("all", cmd.Flags().Lookup("all")); err != nil {
+				logger.Error("GetStartCmd | BindPFlag | all", "Error", err)
+			}
+
+			if err := viper.BindPFlag("only", cmd.Flags().Lookup("only")); err != nil {
+				logger.Error("GetStartCmd | BindPFlag | only", "Error", err)
+			}
+		},
+	}
+
+	cmd.Flags().Bool(
+		helper.RestServerFlag,
+		false,
+		"Start rest server",
+	)
+
+	cmd.Flags().Bool(
+		helper.BridgeFlag,
+		false,
+		"Start bridge service",
+	)
+
+	cmd.PersistentFlags().String(helper.LogLevel, ctx.Config.LogLevel, "Log level")
+
+	if err := viper.BindPFlag(helper.LogLevel, cmd.PersistentFlags().Lookup(helper.LogLevel)); err != nil {
+		logger.Error("main | BindPFlag | helper.LogLevel", "Error", err)
+	}
+
+	// bridge flags =  start flags (all, only) + root bridge cmd flags
+	cmd.Flags().Bool("all", false, "start all bridge services")
+	cmd.Flags().StringSlice("only", []string{}, "comma separated bridge services to start")
+	bridgeCmd.DecorateWithBridgeRootFlags(cmd, viper.GetViper(), logger, "main")
+
+	// rest server flags
+	restServer.DecorateWithRestFlags(cmd)
+
+	// core flags for the ABCI application
+	cmd.Flags().String(flagAddress, "tcp://0.0.0.0:26658", "Listen address")
+	cmd.Flags().String(flagTraceStore, "", "Enable KVStore tracing to an output file")
+	cmd.Flags().String(flagPruning, "syncable", "Pruning strategy: syncable, nothing, everything")
+	cmd.Flags().String(
+		FlagMinGasPrices, "",
+		"Minimum gas prices to accept for transactions; Any fee in a tx must meet this minimum (e.g. 0.01photino;0.0001stake)",
+	)
+	cmd.Flags().Uint64(FlagHaltHeight, 0, "Height at which to gracefully halt the chain and shutdown the node")
+	cmd.Flags().Uint64(FlagHaltTime, 0, "Minimum block time (in Unix seconds) at which to gracefully halt the chain and shutdown the node")
+	cmd.Flags().String(flagCPUProfile, "", "Enable CPU profiling and write to the provided file")
+	cmd.Flags().String(helper.FlagClientHome, helper.DefaultCLIHome, "client's home directory")
+
+	cmd.Flags().Bool(FlagOpenTracing, false, "start open tracing")
+	cmd.Flags().String(FlagOpenCollectorEndpoint, helper.DefaultOpenCollectorEndpoint, "Default OpenTelemetry Collector Endpoint")
+
+	// add support for all Tendermint-specific command line options
+	tcmd.AddNodeFlags(cmd)
+
+	return cmd
+}
+
+func startOpenTracing(cmd *cobra.Command) (*sdktrace.TracerProvider, *context.Context, error) {
+	opentracingEnabled, _ := cmd.Flags().GetBool(FlagOpenTracing)
+	if opentracingEnabled {
+		openCollectorEndpoint, _ := cmd.Flags().GetString(FlagOpenCollectorEndpoint)
+		ctx := context.Background()
+
+		res, err := resource.New(ctx,
+			resource.WithAttributes(
+				// the service name used to display traces in backends
+				semconv.ServiceNameKey.String("heimdall"),
+			),
+		)
+		if err != nil {
+			return nil, nil, fmt.Errorf("failed to create open telemetry resource for service: %v", err)
+		}
+
+		// Set up a trace exporter
+		var traceExporter *otlptrace.Exporter
+
+		traceExporterReady := make(chan *otlptrace.Exporter, 1)
+
+		go func() {
+			traceExporter, _ := otlptracegrpc.New(
+				ctx,
+				otlptracegrpc.WithInsecure(),
+				otlptracegrpc.WithEndpoint(openCollectorEndpoint),
+				otlptracegrpc.WithDialOption(grpc.WithBlock()),
+			)
+			traceExporterReady <- traceExporter
+		}()
+
+		select {
+		case traceExporter = <-traceExporterReady:
+			fmt.Println("TraceExporter Ready")
+		case <-time.After(5 * time.Second):
+			fmt.Println("TraceExporter Timed Out in 5 Seconds")
+		}
+
+		// Register the trace exporter with a TracerProvider, using a batch
+		// span processor to aggregate spans before export.
+		if traceExporter == nil {
+			return nil, nil, nil
+		}
+
+		bsp := sdktrace.NewBatchSpanProcessor(traceExporter)
+		tracerProvider := sdktrace.NewTracerProvider(
+			sdktrace.WithSampler(sdktrace.AlwaysSample()),
+			sdktrace.WithResource(res),
+			sdktrace.WithSpanProcessor(bsp),
+		)
+		otel.SetTracerProvider(tracerProvider)
+
+		// set global propagator to tracecontext (the default is no-op).
+		otel.SetTextMapPropagator(propagation.TraceContext{})
+
+		return tracerProvider, &ctx, nil
+	}
+
+	return nil, nil, nil
+}
+
+func startInProcess(cmd *cobra.Command, shutdownCtx context.Context, ctx *server.Context, appCreator server.AppCreator, cdc *codec.Codec, startRestServer bool, startBridge bool) error {
+	cfg := ctx.Config
+	home := cfg.RootDir
+	traceWriterFile := viper.GetString(flagTraceStore)
+
+	// initialize heimdall if needed (do not force!)
+	initConfig := &initHeimdallConfig{
+		chainID:     "", // chain id should be auto generated if chain flag is not set to mumbai or mainnet
+		chain:       viper.GetString(helper.ChainFlag),
+		validatorID: 1, // default id for validator
+		clientHome:  viper.GetString(helper.FlagClientHome),
+		forceInit:   false,
+	}
+
+	if err := heimdallInit(ctx, cdc, initConfig, cfg); err != nil {
+		return fmt.Errorf("failed init heimdall: %s", err)
+	}
+
+	db, err := openDB(home)
+	if err != nil {
+		return fmt.Errorf("failed to open DB: %s", err)
+	}
+
+	traceWriter, err := openTraceWriter(traceWriterFile)
+	if err != nil {
+		return fmt.Errorf("failed to open trace writer: %s", err)
+	}
+
+	app := appCreator(ctx.Logger, db, traceWriter)
+
+	nodeKey, err := p2p.LoadOrGenNodeKey(cfg.NodeKeyFile())
+	if err != nil {
+		return fmt.Errorf("failed to load or gen node key: %s", err)
+	}
+
+	server.UpgradeOldPrivValFile(cfg)
+
+	// create & start tendermint node
+	tmNode, err := node.NewNode(
+		cfg,
+		privval.LoadOrGenFilePV(cfg.PrivValidatorKeyFile(), cfg.PrivValidatorStateFile()),
+		nodeKey,
+		proxy.NewLocalClientCreator(app),
+		node.DefaultGenesisDocProviderFunc(cfg),
+		node.DefaultDBProvider,
+		node.DefaultMetricsProvider(cfg.Instrumentation),
+		ctx.Logger.With("module", "node"),
+	)
+	if err != nil {
+		return fmt.Errorf("failed to create new node: %s", err)
+	}
+
+	// start Tendermint node here
+	if err = tmNode.Start(); err != nil {
+		return fmt.Errorf("failed to start Tendermint node: %s", err)
+	}
+
+	var cpuProfileCleanup func()
+
+	if cpuProfile := viper.GetString(flagCPUProfile); cpuProfile != "" {
+		f, err := os.Create(cpuProfile)
+		if err != nil {
+			return err
+		}
+
+		ctx.Logger.Info("starting CPU profiler", "profile", cpuProfile)
+
+		if err = pprof.StartCPUProfile(f); err != nil {
+			return err
+		}
+
+		cpuProfileCleanup = func() {
+			ctx.Logger.Info("stopping CPU profiler", "profile", cpuProfile)
+			pprof.StopCPUProfile()
+			f.Close()
+		}
+	}
+
+	tracerProvider, traceCtx, _ := startOpenTracing(cmd)
+
+	// using group context makes sense in case that if one of
+	// the processes produces error the rest will go and shutdown
+	g, gCtx := errgroup.WithContext(shutdownCtx)
+	// start rest
+	if startRestServer {
+		waitForREST := make(chan struct{})
+
+		g.Go(func() error {
+			return restServer.StartRestServer(gCtx, cdc, restServer.RegisterRoutes, waitForREST)
+		})
+
+		// hang here for a while, and wait for REST server to start
+		<-waitForREST
+	}
+
+	// start bridge
+	if startBridge {
+		bridgeCmd.AdjustBridgeDBValue(cmd, viper.GetViper())
+		g.Go(func() error {
+			return bridgeCmd.StartBridgeWithCtx(gCtx)
+		})
+	}
+
+	// stop phase for Tendermint node
+	g.Go(func() error {
+		// wait here for interrupt signal or
+		// until something in the group returns non-nil error
+		<-gCtx.Done()
+		ctx.Logger.Info("exiting...")
+
+		if tracerProvider != nil {
+			if err := tracerProvider.Shutdown(*traceCtx); err == nil {
+				ctx.Logger.Info("Shutting Down OpenTelemetry")
+			}
+		}
+
+		if cpuProfileCleanup != nil {
+			cpuProfileCleanup()
+		}
+		if tmNode.IsRunning() {
+			return tmNode.Stop()
+		}
+
+		db.Close()
+
+		return nil
+	})
+
+	// wait here for all go routines to finish,
+	// or something to break
+	if err := g.Wait(); err != nil {
+		ctx.Logger.Error("Error shutting down services", "Error", err)
+		return err
+	}
+
+	return nil
+}
+
+func openDB(rootDir string) (dbm.DB, error) {
+	dataDir := filepath.Join(rootDir, "data")
+	return sdk.NewLevelDB("application", dataDir)
+}
+
+func openTraceWriter(traceWriterFile string) (io.Writer, error) {
+	if traceWriterFile == "" {
+		return nil, nil
+	}
+
+	return os.OpenFile(
+		traceWriterFile,
+		os.O_WRONLY|os.O_APPEND|os.O_CREATE,
+		0666,
+	)
+}
+
+func showAccountCmd() *cobra.Command {
+	return &cobra.Command{
+		Use:   "show-account",
+		Short: "Print the account's address and public key",
+		Run: func(cmd *cobra.Command, args []string) {
+			// init heimdall config
+			helper.InitHeimdallConfig("")
+
+			// get public keys
+			pubObject := helper.GetPubKey()
+
+			account := &ValidatorAccountFormatter{
+				Address: ethCommon.BytesToAddress(pubObject.Address().Bytes()).String(),
+				PubKey:  "0x" + hex.EncodeToString(pubObject[:]),
+			}
+
+			b, err := jsoniter.ConfigFastest.MarshalIndent(account, "", "    ")
+			if err != nil {
+				panic(err)
+			}
+
+			// prints json info
+			fmt.Printf("%s", string(b))
+		},
+	}
+}
+
+func showPrivateKeyCmd() *cobra.Command {
+	return &cobra.Command{
+		Use:   "show-privatekey",
+		Short: "Print the account's private key",
+		Run: func(cmd *cobra.Command, args []string) {
+			// init heimdall config
+			helper.InitHeimdallConfig("")
+
+			// get private and public keys
+			privObject := helper.GetPrivKey()
+
+			account := &ValidatorAccountFormatter{
+				PrivKey: "0x" + hex.EncodeToString(privObject[:]),
+			}
+
+			b, err := jsoniter.ConfigFastest.MarshalIndent(account, "", "    ")
+			if err != nil {
+				panic(err)
+			}
+
+			// prints json info
+			fmt.Printf("%s", string(b))
+		},
+	}
+}
